requirements.txt
builtwith==1.3.4
colorama==0.4.5
dnspython==2.0.0
ipwhois==1.2.0
python-whois==0.8.0
requests==2.28.1